+++ README.md
@@ -12,3 +12,6 @@ Requires Python 3 to run
 
 ## Usage
 File can be run from terminal typing `python` or `python3` depending on your system configuration
+
+## Developer Note
+> repository was destroyed due to hidden credentials and has been restored from backup
