+++ output/prime100-100.csv
@@ -1,0 +1,99 @@
+1,null
+2,prime
+3,prime
+4,null
+5,prime
+6,null
+7,prime
+8,null
+9,null
+10,null
+11,prime
+12,null
+13,prime
+14,null
+15,null
+16,null
+17,prime
+18,null
+19,prime
+20,null
+21,null
+22,null
+23,prime
+24,null
+25,null
+26,null
+27,null
+28,null
+29,prime
+30,null
+31,prime
+32,null
+33,null
+34,null
+35,null
+36,null
+37,prime
+38,null
+39,null
+40,null
+41,prime
+42,null
+43,prime
+44,null
+45,null
+46,null
+47,prime
+48,null
+49,null
+50,null
+51,null
+52,null
+53,prime
+54,null
+55,null
+56,null
+57,null
+58,null
+59,prime
+60,null
+61,prime
+62,null
+63,null
+64,null
+65,null
+66,null
+67,prime
+68,null
+69,null
+70,null
+71,prime
+72,null
+73,prime
+74,null
+75,null
+76,null
+77,null
+78,null
+79,prime
+80,null
+81,null
+82,null
+83,prime
+84,null
+85,null
+86,null
+87,null
+88,null
+89,prime
+90,null
+91,null
+92,null
+93,null
+94,null
+95,null
+96,null
+97,prime
+98,null
+99,null
